Implement Windows ssh
