+++ src/main/java/hudson/plugins/ec2/WindowsSSHData.java
@@ -5,24 +5,26 @@ import hudson.Extension;
 import hudson.model.Descriptor;
 import org.kohsuke.stapler.DataBoundConstructor;
 
-public class UnixData extends SSHData {
+
+public class WindowsSSHData extends SSHData {
 
     @DataBoundConstructor
-    public UnixData(String rootCommandPrefix, String slaveCommandPrefix, String slaveCommandSuffix, String sshPort, String bootDelay) {
+    public WindowsSSHData(String rootCommandPrefix, String slaveCommandPrefix, String slaveCommandSuffix, String sshPort, String bootDelay) {
         super(rootCommandPrefix, slaveCommandPrefix, slaveCommandSuffix, sshPort, bootDelay);
     }
 
     @Override
-    public boolean isUnix() {
+    public boolean isWindowsSSH() {
         return true;
     }
+
 
     @Extension
     public static class DescriptorImpl extends Descriptor<AMITypeData> {
         @Override
         @NonNull
         public String getDisplayName() {
-            return "unix";
+            return "windows-ssh";
         }
     }
 }
