+++ src/main/resources/hudson/plugins/ec2/WindowsSSHData/help-bootDelay.html
@@ -1,0 +1,6 @@
+<div>
+Indicate here the time in seconds to wait for the machine to be ready once the plugin detects SSH is available.
+Unfortunately, on Windows during the boot, the SSH service might be started, and then several minutes after will be
+restarted. If this restart happens during the slave provisioning, Windows will prevent subsequent SSH connections and
+the slave will not be correctly provisioned.
+</div>
